# backend/api/openapi.yaml
openapi: "3.0.3"
info:
  title: eco-base API
  version: "1.0.0"
  description: |
    Enterprise cloud-native AI platform API.
    URI: eco-base://backend/api/openapi

    Provides authentication, platform management, AI inference (OpenAI-compatible),
    vector alignment, async job management, embedding, and .qyaml governance.
  contact:
    name: eco-base Platform Team
    email: platform-team@autoecoops.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development (API gateway)
  - url: http://localhost:8001
    description: Local development (AI service)
  - url: https://api.autoecoops.io
    description: Production
...
paths:
...
  /auth/signup:
    post:
      summary: Create account
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
                display_name: { type: string }
      responses:
        "201":
          description: Account created
        "409":
          description: Email already registered
  /auth/login:
    post:
      summary: Sign in
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful, returns session tokens
        "401":
          description: Invalid credentials
  /auth/refresh:
    post:
      summary: Refresh session token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New session tokens
        "401":
          description: Invalid or expired refresh token
  /auth/logout:
    post:
      summary: Sign out
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
  /auth/me:
    get:
      summary: Get current user profile
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
        "401":
          description: Not authenticated
  /api/v1/platforms:
    get:
      summary: List platforms
      tags: [Platforms]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform list
    post:
      summary: Register platform (admin only)
      tags: [Platforms]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
                capabilities: { type: array, items: { type: string } }
                deployTarget: { type: string }
      responses:
        "201":
          description: Platform created
        "403":
          description: Admin access required
...
  /v1/chat/completions:
    post:
      summary: OpenAI-compatible chat completions
      tags: [OpenAI]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Chat completion response (or SSE stream if stream=true)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
            text/event-stream:
              schema:
                type: string
        "422":
          description: Validation error
...
  /v1/models:
    get:
      summary: OpenAI-compatible model list
      tags: [OpenAI]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Model list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelListResponse"
...
  /api/v1/jobs:
    post:
      summary: Submit async inference job
      tags: [Jobs]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string }
                model_id: { type: string, default: "default" }
                max_tokens: { type: integer, default: 2048 }
                priority: { type: string, enum: [high, normal, low], default: normal }
      responses:
        "202":
          description: Job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string, format: uuid }
                  status: { type: string, example: pending }
    get:
      summary: List inference jobs
      tags: [Jobs]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, running, completed, failed, cancelled] }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: Job list
  /api/v1/jobs/{jobId}:
    get:
      summary: Get job status
      tags: [Jobs]
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Job details
        "404":
          description: Job not found
    delete:
      summary: Cancel job
      tags: [Jobs]
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Job cancelled
        "404":
          description: Job not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChatCompletionRequest:
      type: object
      required: [messages]
      properties:
        model: { type: string, default: "default" }
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role: { type: string, enum: [system, user, assistant] }
              content: { type: string }
        temperature: { type: number, minimum: 0, maximum: 2 }
        max_tokens: { type: integer, default: 2048 }
        stream: { type: boolean, default: false }
    ChatCompletionResponse:
      type: object
      properties:
        id: { type: string }
        object: { type: string, example: chat.completion }
        model: { type: string }
        choices:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              message:
                type: object
                properties:
                  role: { type: string }
                  content: { type: string }
              finish_reason: { type: string }
        usage:
          $ref: "#/components/schemas/UsageInfo"
        system_fingerprint: { type: string }
...
    ModelListResponse:
      type: object
      properties:
        object: { type: string, example: list }
        data:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              object: { type: string, example: model }
              owned_by: { type: string, example: eco-base }
              capabilities: { type: array, items: { type: string } }
              compatible_engines: { type: array, items: { type: string } }
              context_length: { type: integer }
...
    UsageInfo:
      type: object
      properties:
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }